tests/test_cache.py
"""Tests for caching modules (CAG and Semantic Cache)."""

import pytest
import numpy as np
from datetime import datetime, timedelta

from perfect_rag.cache.semantic_cache import SemanticCache, SemanticCacheEntry


class TestSemanticCache:
    """Tests for semantic query cache."""

    @pytest.fixture
    def semantic_cache(self, mock_embedding_service, settings):
        """Create semantic cache for testing."""
        return SemanticCache(
            embedding_service=mock_embedding_service,
            settings=settings,
        )

    @pytest.mark.asyncio
    async def test_store_and_exact_lookup(self, semantic_cache, mock_embedding_service):
        """Test storing and retrieving with exact match."""
        query = "What is machine learning?"
        embedding = await mock_embedding_service.embed_query(query)

        # Store entry
        await semantic_cache.store(
            query=query,
            query_embedding=embedding,
            response="Machine learning is a subset of AI...",
            citations=[{"source": "ml_guide.pdf"}],
            confidence=0.95,
            model="test-model",
        )

        # Lookup with same query
        entry, similarity = await semantic_cache.lookup(query, embedding)

        assert entry is not None
        assert similarity == 1.0  # Exact match
        assert entry.query == query
        assert entry.response == "Machine learning is a subset of AI..."

    @pytest.mark.asyncio
    async def test_similar_query_lookup(self, semantic_cache, mock_embedding_service):
        """Test retrieving with similar query."""
        original_query = "What is machine learning?"
        original_embedding = await mock_embedding_service.embed_query(original_query)

        await semantic_cache.store(
            query=original_query,
            query_embedding=original_embedding,
            response="ML is a field of AI...",
            confidence=0.9,
        )

        # Very similar query (same hash with mock embedder)
        similar_query = "what is machine learning?"  # lowercase
        similar_embedding = await mock_embedding_service.embed_query(similar_query)

        entry, similarity = await semantic_cache.lookup(similar_query, similar_embedding)

        # Should find as similar (depending on similarity threshold)
        if entry is not None:
            assert similarity >= semantic_cache.similar_threshold

    @pytest.mark.asyncio
    async def test_no_match_for_different_query(self, semantic_cache, mock_embedding_service):
        """Test no match for unrelated query."""
        await semantic_cache.store(
            query="What is machine learning?",
            query_embedding=await mock_embedding_service.embed_query("What is machine learning?"),
            response="ML is...",
            confidence=0.9,
        )

        # Completely different query
        different_query = "How do I cook pasta?"
        different_embedding = await mock_embedding_service.embed_query(different_query)

        entry, similarity = await semantic_cache.lookup(different_query, different_embedding)

        # Should not find a match (or low similarity)
        if entry is not None:
            assert similarity < semantic_cache.similar_threshold

    @pytest.mark.asyncio
    async def test_access_count_increments(self, semantic_cache, mock_embedding_service):
        """Test that access count increments on lookup."""
        query = "Test query"
        embedding = await mock_embedding_service.embed_query(query)

        await semantic_cache.store(
            query=query,
            query_embedding=embedding,
            response="Test response",
            confidence=0.9,
        )

        # First lookup
        entry1, _ = await semantic_cache.lookup(query, embedding)
        initial_count = entry1.access_count

        # Second lookup
        entry2, _ = await semantic_cache.lookup(query, embedding)

        assert entry2.access_count == initial_count + 1

    @pytest.mark.asyncio
    async def test_ttl_expiration(self, semantic_cache, mock_embedding_service):
        """Test that entries expire after TTL."""
        query = "Expiring query"
        embedding = await mock_embedding_service.embed_query(query)

        # Store with very short TTL
        entry = await semantic_cache.store(
            query=query,
            query_embedding=embedding,
            response="Expiring response",
            confidence=0.9,
            ttl_minutes=0,  # Immediate expiration
        )

        # Manually set TTL to past
        entry.ttl = datetime.utcnow() - timedelta(minutes=1)

        # Should not find expired entry
        result, _ = await semantic_cache.lookup(query, embedding)
        # Note: The exact match lookup might still work since we check TTL differently
        # The semantic match should filter it out

    @pytest.mark.asyncio
    async def test_invalidate_by_content(self, semantic_cache, mock_embedding_service):
        """Test invalidating entries by content."""
        query1 = "Query about Python"
        query2 = "Query about JavaScript"

        await semantic_cache.store(
            query=query1,
            query_embedding=await mock_embedding_service.embed_query(query1),
            response="Python is a programming language",
            confidence=0.9,
        )

        await semantic_cache.store(
            query=query2,
            query_embedding=await mock_embedding_service.embed_query(query2),
            response="JavaScript is used for web development",
            confidence=0.9,
        )

        # Invalidate entries containing "Python"
        count = await semantic_cache.invalidate_by_content("Python")

        assert count == 1

    @pytest.mark.asyncio
    async def test_get_stats(self, semantic_cache, mock_embedding_service):
        """Test getting cache statistics."""
        await semantic_cache.store(
            query="Test",
            query_embedding=await mock_embedding_service.embed_query("Test"),
            response="Response",
            confidence=0.9,
        )

        stats = await semantic_cache.get_stats()

        assert "total_entries" in stats
        assert "active_entries" in stats
        assert "max_entries" in stats
        assert stats["total_entries"] == 1

    @pytest.mark.asyncio
    async def test_clear(self, semantic_cache, mock_embedding_service):
        """Test clearing the cache."""
        # Add some entries
        for i in range(3):
            await semantic_cache.store(
                query=f"Query {i}",
                query_embedding=await mock_embedding_service.embed_query(f"Query {i}"),
                response=f"Response {i}",
                confidence=0.9,
            )

        # Clear
        count = await semantic_cache.clear()

        assert count == 3

        stats = await semantic_cache.get_stats()
        assert stats["total_entries"] == 0

    def test_adapt_response_high_similarity(self, semantic_cache):
        """Test response adaptation for high similarity."""
        entry = SemanticCacheEntry(
            key="test",
            query="Original query",
            query_embedding=[0.1] * 1024,
            response="Original response",
            citations=[],
            confidence=0.9,
            model="test",
        )

        # High similarity - should return as-is
        adapted = semantic_cache.adapt_response(entry, "Very similar query", 0.99)
        assert adapted == entry.response

    def test_adapt_response_lower_similarity(self, semantic_cache):
        """Test response adaptation for lower similarity."""
        entry = SemanticCacheEntry(
            key="test",
            query="Original query",
            query_embedding=[0.1] * 1024,
            response="Original response",
            citations=[],
            confidence=0.9,
            model="test",
        )

        # Lower similarity - currently returns same but could add metadata
        adapted = semantic_cache.adapt_response(entry, "Somewhat similar query", 0.90)
        assert adapted == entry.response


class TestEviction:
    """Tests for cache eviction."""

    @pytest.mark.asyncio
    async def test_eviction_when_full(self, mock_embedding_service, settings):
        """Test eviction when cache is full."""
        cache = SemanticCache(
            embedding_service=mock_embedding_service,
            settings=settings,
        )
        cache.max_entries = 5  # Small cache for testing

        # Fill the cache
        for i in range(6):
            await cache.store(
                query=f"Query {i}",
                query_embedding=await mock_embedding_service.embed_query(f"Query {i}"),
                response=f"Response {i}",
                confidence=0.9,
            )

        # Should have evicted some entries
        stats = await cache.get_stats()
        assert stats["total_entries"] <= cache.max_entries
