pyproject.toml
[project]
name = "perfect-rag"
version = "0.1.0"
description = "Production-ready RAG system with GraphRAG, hybrid search, and OpenAI-compatible API"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Perfect RAG Team" }]
keywords = ["rag", "graphrag", "llm", "embeddings", "knowledge-graph"]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1.0",

    # Data validation & settings
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Database clients
    "surrealdb>=0.3.0",
    "qdrant-client>=1.12.0",
    "httpx>=0.27.0",

    # Embeddings & ML
    "sentence-transformers>=3.3.0",
    "torch>=2.4.0",
    "transformers>=4.46.0",
    "tiktoken>=0.8.0",

    # LLM providers
    "openai>=1.54.0",
    "anthropic>=0.39.0",

    # Document processing
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "pygments>=2.18.0",

    # NLP
    "spacy>=3.8.0",

    # Utilities
    "aiofiles>=24.1.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "rdflib>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

ocr = [
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
]

multimodal = [
    # ColPali for vision-language retrieval
    "colpali-engine>=0.3.0",
    # OCR dependencies
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    # Image processing
    "pillow>=10.0.0",
]

colbert = [
    # ColBERT / Late Interaction (requires sentence-transformers<3.0.0)
    # Install separately: pip install ragatouille
    # Note: May require downgrading sentence-transformers
]

gpu = [
    # Install CUDA PyTorch separately: pip install torch --index-url https://download.pytorch.org/whl/cu124
]

[project.scripts]
rag = "perfect_rag.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/perfect_rag"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
